Read EPP fitting surveillance data

Reads the HIV prevalence from sentinel surveillance and household surveillance data from the EPP .xml file within a PJNZ.

read_epp_data(pjnz_path)

Arguments

pjnz_path
Path to Spectrum PJNZ file.

Value

List of data frames containing prevalence and incidence data.

Details

EPP projection sets are identified in the .xml file by searching the XML tree for tag "object", and then selecting objects with "class" attribute equal to "epp2011.core.sets.ProjectionSet".

Examples

pjnz_path <- system.file("testdata", "Botswana2018.PJNZ", package = "specio")
read_epp_data(pjnz_path)

#> $Urban
#> $Urban$country
#> [1] "Botswana"
#> 
#> $Urban$region
#> [1] "Urban"
#> 
#> $Urban$projset_id
#> [1] 0
#> 
#> $Urban$anc.used
#>  [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> 
#> $Urban$anc.prev
#>                     year
#> site                 1985 1986 1987 1988 1989 1990 1991  1992  1993  1994  1995
#>   Gaborone (%)         NA   NA   NA   NA   NA   NA 0.17 0.149 0.192 0.278 0.287
#>   Francistown (%)      NA   NA   NA   NA   NA   NA 0.16 0.237 0.342 0.297 0.396
#>   Southern (%)         NA   NA   NA   NA   NA   NA   NA    NA    NA 0.160    NA
#>   South East (%)       NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>   Kweneng East (%)     NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>   Mahalapye (%)        NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>   Serowe Palapye (%)   NA   NA   NA   NA   NA   NA   NA    NA 0.199    NA 0.299
#>   Lobatse (%)          NA   NA   NA   NA   NA   NA   NA    NA 0.178    NA 0.389
#>   Selebi-Pikwe (%)     NA   NA   NA   NA   NA   NA   NA    NA    NA 0.270    NA
#>   Jwaneng (%)          NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>                     year
#> site                  1996  1997  1998  1999  2000  2001  2002  2003 2004  2005
#>   Gaborone (%)       0.313 0.340 0.391 0.371 0.362 0.386 0.382 0.448   NA 0.344
#>   Francistown (%)    0.431 0.429 0.430 0.427 0.444 0.496 0.402 0.458   NA 0.423
#>   Southern (%)       0.217    NA 0.247    NA 0.407 0.340 0.331 0.257   NA 0.282
#>   South East (%)        NA    NA    NA    NA    NA 0.321 0.265 0.279   NA 0.321
#>   Kweneng East (%)   0.239    NA 0.372    NA 0.304 0.296 0.292 0.321   NA 0.315
#>   Mahalapye (%)         NA 0.282    NA 0.320    NA 0.319 0.398 0.374   NA 0.362
#>   Serowe Palapye (%)    NA 0.344    NA 0.418    NA 0.446 0.367 0.433   NA 0.375
#>   Lobatse (%)           NA 0.337    NA 0.313    NA 0.306 0.346 0.324   NA 0.290
#>   Selebi-Pikwe (%)   0.378    NA 0.499    NA 0.503 0.500 0.481 0.522   NA 0.465
#>   Jwaneng (%)           NA    NA    NA    NA    NA    NA    NA    NA   NA 0.349
#>                     year
#> site                  2006  2007 2008  2009 2010  2011 2012 2013 2014 2015 2016
#>   Gaborone (%)       0.353 0.332   NA 0.290   NA 0.292   NA   NA   NA   NA   NA
#>   Francistown (%)    0.367 0.379   NA 0.381   NA 0.370   NA   NA   NA   NA   NA
#>   Southern (%)       0.250 0.228   NA 0.283   NA 0.224   NA   NA   NA   NA   NA
#>   South East (%)     0.276 0.267   NA 0.288   NA 0.158   NA   NA   NA   NA   NA
#>   Kweneng East (%)   0.317 0.335   NA 0.279   NA 0.266   NA   NA   NA   NA   NA
#>   Mahalapye (%)      0.299 0.304   NA 0.378   NA 0.335   NA   NA   NA   NA   NA
#>   Serowe Palapye (%) 0.368 0.373   NA 0.336   NA 0.379   NA   NA   NA   NA   NA
#>   Lobatse (%)        0.284 0.287   NA 0.324   NA 0.244   NA   NA   NA   NA   NA
#>   Selebi-Pikwe (%)   0.411 0.490   NA 0.391   NA 0.412   NA   NA   NA   NA   NA
#>   Jwaneng (%)        0.200 0.214   NA 0.311   NA 0.167   NA   NA   NA   NA   NA
#>                     year
#> site                 2017 2018 2019 2020 2021 2022 2023 2024 2025
#>   Gaborone (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Francistown (%)      NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Southern (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   South East (%)       NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng East (%)     NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Mahalapye (%)        NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Serowe Palapye (%)   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Lobatse (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Selebi-Pikwe (%)     NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Jwaneng (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $Urban$anc.n
#>                     year
#> site                 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995
#>   Gaborone (%)         NA   NA   NA   NA   NA   NA   58  841  881 1205 1307
#>   Francistown (%)      NA   NA   NA   NA   NA   NA  128  796  803  799  626
#>   Southern (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA  508   NA
#>   South East (%)       NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng East (%)     NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Mahalapye (%)        NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Serowe Palapye (%)   NA   NA   NA   NA   NA   NA   NA   NA  267   NA  262
#>   Lobatse (%)          NA   NA   NA   NA   NA   NA   NA   NA  258   NA  231
#>   Selebi-Pikwe (%)     NA   NA   NA   NA   NA   NA   NA   NA   NA  307   NA
#>   Jwaneng (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>                     year
#> site                 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006
#>   Gaborone (%)       1232  786 1251  480  734  576  666  555   NA  722  607
#>   Francistown (%)     751  802  796  576  702  494  458  593   NA  659  560
#>   Southern (%)        589   NA  300   NA  513  347  363  359   NA  313  324
#>   South East (%)       NA   NA   NA   NA   NA  192  288  306   NA  368  385
#>   Kweneng East (%)    213   NA  534   NA  349  540  567  522   NA  551  518
#>   Mahalapye (%)        NA  393   NA  362   NA  336  303  326   NA  379  322
#>   Serowe Palapye (%)   NA  289   NA  268   NA  341  425  389   NA  534  598
#>   Lobatse (%)          NA  266   NA  252   NA  232  241  246   NA  243  249
#>   Selebi-Pikwe (%)    331   NA  471   NA  304  351  372  302   NA  300  324
#>   Jwaneng (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA  146  143
#>                     year
#> site                 2007 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017
#>   Gaborone (%)        692   NA  678   NA  572   NA   NA   NA   NA   NA   NA
#>   Francistown (%)     702   NA  527   NA  547   NA   NA   NA   NA   NA   NA
#>   Southern (%)        319   NA  316   NA  302   NA   NA   NA   NA   NA   NA
#>   South East (%)      441   NA  421   NA  242   NA   NA   NA   NA   NA   NA
#>   Kweneng East (%)    487   NA  486   NA  262   NA   NA   NA   NA   NA   NA
#>   Mahalapye (%)       403   NA  409   NA  463   NA   NA   NA   NA   NA   NA
#>   Serowe Palapye (%)  667   NA  405   NA  458   NA   NA   NA   NA   NA   NA
#>   Lobatse (%)         242   NA  274   NA  169   NA   NA   NA   NA   NA   NA
#>   Selebi-Pikwe (%)    310   NA  263   NA  282   NA   NA   NA   NA   NA   NA
#>   Jwaneng (%)         141   NA  300   NA  300   NA   NA   NA   NA   NA   NA
#>                     year
#> site                 2018 2019 2020 2021 2022 2023 2024 2025
#>   Gaborone (%)         NA   NA   NA   NA   NA   NA   NA   NA
#>   Francistown (%)      NA   NA   NA   NA   NA   NA   NA   NA
#>   Southern (%)         NA   NA   NA   NA   NA   NA   NA   NA
#>   South East (%)       NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng East (%)     NA   NA   NA   NA   NA   NA   NA   NA
#>   Mahalapye (%)        NA   NA   NA   NA   NA   NA   NA   NA
#>   Serowe Palapye (%)   NA   NA   NA   NA   NA   NA   NA   NA
#>   Lobatse (%)          NA   NA   NA   NA   NA   NA   NA   NA
#>   Selebi-Pikwe (%)     NA   NA   NA   NA   NA   NA   NA   NA
#>   Jwaneng (%)          NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $Urban$ancrtsite.prev
#>                     year
#> site                 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995
#>   Gaborone (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Francistown (%)      NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Southern (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   South East (%)       NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng East (%)     NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Mahalapye (%)        NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Serowe Palapye (%)   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Lobatse (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Selebi-Pikwe (%)     NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Jwaneng (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>                     year
#> site                 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006
#>   Gaborone (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Francistown (%)      NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Southern (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   South East (%)       NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng East (%)     NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Mahalapye (%)        NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Serowe Palapye (%)   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Lobatse (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Selebi-Pikwe (%)     NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Jwaneng (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>                     year
#> site                 2007 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017
#>   Gaborone (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Francistown (%)      NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Southern (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   South East (%)       NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng East (%)     NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Mahalapye (%)        NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Serowe Palapye (%)   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Lobatse (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Selebi-Pikwe (%)     NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Jwaneng (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>                     year
#> site                 2018 2019 2020 2021 2022 2023 2024 2025
#>   Gaborone (%)         NA   NA   NA   NA   NA   NA   NA   NA
#>   Francistown (%)      NA   NA   NA   NA   NA   NA   NA   NA
#>   Southern (%)         NA   NA   NA   NA   NA   NA   NA   NA
#>   South East (%)       NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng East (%)     NA   NA   NA   NA   NA   NA   NA   NA
#>   Mahalapye (%)        NA   NA   NA   NA   NA   NA   NA   NA
#>   Serowe Palapye (%)   NA   NA   NA   NA   NA   NA   NA   NA
#>   Lobatse (%)          NA   NA   NA   NA   NA   NA   NA   NA
#>   Selebi-Pikwe (%)     NA   NA   NA   NA   NA   NA   NA   NA
#>   Jwaneng (%)          NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $Urban$ancrtsite.n
#>                     year
#> site                 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995
#>   Gaborone (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Francistown (%)      NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Southern (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   South East (%)       NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng East (%)     NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Mahalapye (%)        NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Serowe Palapye (%)   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Lobatse (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Selebi-Pikwe (%)     NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Jwaneng (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>                     year
#> site                 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006
#>   Gaborone (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Francistown (%)      NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Southern (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   South East (%)       NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng East (%)     NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Mahalapye (%)        NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Serowe Palapye (%)   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Lobatse (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Selebi-Pikwe (%)     NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Jwaneng (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>                     year
#> site                 2007 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017
#>   Gaborone (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Francistown (%)      NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Southern (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   South East (%)       NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng East (%)     NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Mahalapye (%)        NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Serowe Palapye (%)   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Lobatse (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Selebi-Pikwe (%)     NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Jwaneng (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>                     year
#> site                 2018 2019 2020 2021 2022 2023 2024 2025
#>   Gaborone (%)         NA   NA   NA   NA   NA   NA   NA   NA
#>   Francistown (%)      NA   NA   NA   NA   NA   NA   NA   NA
#>   Southern (%)         NA   NA   NA   NA   NA   NA   NA   NA
#>   South East (%)       NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng East (%)     NA   NA   NA   NA   NA   NA   NA   NA
#>   Mahalapye (%)        NA   NA   NA   NA   NA   NA   NA   NA
#>   Serowe Palapye (%)   NA   NA   NA   NA   NA   NA   NA   NA
#>   Lobatse (%)          NA   NA   NA   NA   NA   NA   NA   NA
#>   Selebi-Pikwe (%)     NA   NA   NA   NA   NA   NA   NA   NA
#>   Jwaneng (%)          NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $Urban$ancrtcens
#>      year   prev     n
#> 2014 2014 0.2870 25690
#> 2015 2015 0.2800 25467
#> 2016 2016 0.2870 23939
#> 2017 2017 0.2821 26821
#> 
#> $Urban$hhs
#>   year  prev     se  n used incid incid_se prev_incid_corr incid_cohort
#> 1 2008 0.239 0.0110 NA TRUE    NA       NA              NA           NA
#> 2 2004 0.242 0.0108 NA TRUE    NA       NA              NA           NA
#> 3 2012 0.215 0.0100 NA TRUE    NA       NA              NA           NA
#> 
#> 
#> $Rural
#> $Rural$country
#> [1] "Botswana"
#> 
#> $Rural$region
#> [1] "Rural"
#> 
#> $Rural$projset_id
#> [1] 1
#> 
#> $Rural$anc.used
#>  [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> 
#> $Rural$anc.prev
#>                       year
#> site                   1985 1986 1987 1988 1989 1990 1991  1992  1993  1994
#>   North East (%)         NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>   Bobirwa district (%)   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>   Boteti (%)             NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>   Gantsi (%)             NA   NA   NA   NA   NA   NA   NA    NA 0.095    NA
#>   Chobe (%)              NA   NA   NA   NA   NA   NA   NA    NA 0.183    NA
#>   Kgalagadi (%)          NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>   Tutume (%)             NA   NA   NA   NA   NA   NA   NA    NA    NA 0.231
#>   Kweneng West (%)       NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>   Goodhope (%)           NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>   Okavango (%)           NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>   Hukuntsi (%)           NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>   Ngami (%)              NA   NA   NA   NA   NA   NA 0.08 0.127    NA 0.194
#>   Kgatleng (%)           NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>   Mabutsane (%)          NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>                       year
#> site                    1995  1996  1997  1998  1999  2000  2001  2002  2003
#>   North East (%)          NA    NA    NA    NA    NA    NA 0.400 0.424 0.404
#>   Bobirwa district (%)    NA    NA    NA    NA    NA    NA 0.443 0.453 0.493
#>   Boteti (%)              NA    NA    NA    NA    NA    NA 0.367 0.356 0.319
#>   Gantsi (%)           0.189    NA    NA 0.223    NA    NA 0.218 0.188    NA
#>   Chobe (%)            0.379    NA 0.383    NA 0.508    NA 0.383 0.429 0.470
#>   Kgalagadi (%)           NA    NA    NA    NA 0.218    NA 0.285 0.283 0.289
#>   Tutume (%)              NA 0.332    NA 0.375    NA 0.354 0.511 0.407 0.377
#>   Kweneng West (%)        NA 0.301    NA 0.265    NA    NA 0.253 0.287 0.270
#>   Goodhope (%)            NA    NA    NA    NA    NA    NA 0.331 0.263 0.409
#>   Okavango (%)            NA    NA    NA    NA    NA    NA 0.406 0.342 0.327
#>   Hukuntsi (%)            NA    NA    NA    NA    NA    NA 0.233 0.400 0.284
#>   Ngami (%)               NA 0.323    NA 0.335    NA    NA 0.358 0.407 0.384
#>   Kgatleng (%)            NA    NA 0.305    NA 0.295    NA 0.249 0.309    NA
#>   Mabutsane (%)           NA    NA    NA    NA    NA    NA    NA    NA    NA
#>                       year
#> site                   2004  2005  2006  2007 2008  2009 2010  2011 2012 2013
#>   North East (%)         NA 0.349 0.354 0.388   NA 0.365   NA 0.296   NA   NA
#>   Bobirwa district (%)   NA 0.298 0.376 0.429   NA 0.466   NA 0.418   NA   NA
#>   Boteti (%)             NA 0.354 0.290 0.337   NA 0.410   NA 0.274   NA   NA
#>   Gantsi (%)             NA 0.209 0.270 0.266   NA 0.197   NA 0.173   NA   NA
#>   Chobe (%)              NA    NA 0.420 0.456   NA 0.408   NA 0.371   NA   NA
#>   Kgalagadi (%)          NA 0.262 0.191 0.310   NA 0.185   NA 0.272   NA   NA
#>   Tutume (%)             NA 0.413 0.399 0.387   NA 0.363   NA 0.359   NA   NA
#>   Kweneng West (%)       NA 0.270 0.281 0.239   NA 0.288   NA 0.330   NA   NA
#>   Goodhope (%)           NA 0.208 0.230 0.265   NA 0.278   NA 0.308   NA   NA
#>   Okavango (%)           NA 0.292 0.291 0.347   NA 0.333   NA 0.262   NA   NA
#>   Hukuntsi (%)           NA 0.294 0.294 0.327   NA 0.143   NA 0.305   NA   NA
#>   Ngami (%)              NA 0.273 0.354 0.397   NA 0.303   NA 0.278   NA   NA
#>   Kgatleng (%)           NA 0.301 0.243 0.308   NA 0.220   NA 0.345   NA   NA
#>   Mabutsane (%)          NA    NA 0.264 0.253   NA 0.292   NA 0.367   NA   NA
#>                       year
#> site                   2014 2015 2016 2017 2018 2019 2020 2021 2022 2023 2024
#>   North East (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Bobirwa district (%)   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Boteti (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Gantsi (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Chobe (%)              NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgalagadi (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Tutume (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng West (%)       NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Goodhope (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Okavango (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Hukuntsi (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Ngami (%)              NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgatleng (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Mabutsane (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>                       year
#> site                   2025
#>   North East (%)         NA
#>   Bobirwa district (%)   NA
#>   Boteti (%)             NA
#>   Gantsi (%)             NA
#>   Chobe (%)              NA
#>   Kgalagadi (%)          NA
#>   Tutume (%)             NA
#>   Kweneng West (%)       NA
#>   Goodhope (%)           NA
#>   Okavango (%)           NA
#>   Hukuntsi (%)           NA
#>   Ngami (%)              NA
#>   Kgatleng (%)           NA
#>   Mabutsane (%)          NA
#> 
#> $Rural$anc.n
#>                       year
#> site                   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995
#>   North East (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Bobirwa district (%)   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Boteti (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Gantsi (%)             NA   NA   NA   NA   NA   NA   NA   NA  126   NA   95
#>   Chobe (%)              NA   NA   NA   NA   NA   NA   NA   NA   71   NA  103
#>   Kgalagadi (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Tutume (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA  303   NA
#>   Kweneng West (%)       NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Goodhope (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Okavango (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Hukuntsi (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Ngami (%)              NA   NA   NA   NA   NA   NA   53  299   NA  227   NA
#>   Kgatleng (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Mabutsane (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>                       year
#> site                   1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006
#>   North East (%)         NA   NA   NA   NA   NA   75  196  222   NA  213  244
#>   Bobirwa district (%)   NA   NA   NA   NA   NA  149  286  255   NA  209  292
#>   Boteti (%)             NA   NA   NA   NA   NA  281  250  260   NA  331  310
#>   Gantsi (%)             NA   NA  139   NA   NA  159  112   NA   NA  188  201
#>   Chobe (%)              NA   63   NA  120   NA   81  105  115   NA   NA  122
#>   Kgalagadi (%)          NA   NA   NA  257   NA  120  172  161   NA  164  188
#>   Tutume (%)            339   NA  299   NA  339  267  370  429   NA  278  388
#>   Kweneng West (%)       72   NA  136   NA   NA  130  158  190   NA  178  207
#>   Goodhope (%)           NA   NA   NA   NA   NA   96  146  137   NA  184  170
#>   Okavango (%)           NA   NA   NA   NA   NA  225  284  328   NA  288  242
#>   Hukuntsi (%)           NA   NA   NA   NA   NA   62  182  100   NA  118  112
#>   Ngami (%)             331   NA  337   NA   NA  376  253  376   NA  353  428
#>   Kgatleng (%)           NA  294   NA  271   NA  231  310   NA   NA  398  373
#>   Mabutsane (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  112
#>                       year
#> site                   2007 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017
#>   North East (%)        245   NA  226   NA  198   NA   NA   NA   NA   NA   NA
#>   Bobirwa district (%)  262   NA  284   NA  363   NA   NA   NA   NA   NA   NA
#>   Boteti (%)            292   NA  293   NA  197   NA   NA   NA   NA   NA   NA
#>   Gantsi (%)            188   NA  244   NA  135   NA   NA   NA   NA   NA   NA
#>   Chobe (%)             156   NA  138   NA  143   NA   NA   NA   NA   NA   NA
#>   Kgalagadi (%)         164   NA  188   NA  191   NA   NA   NA   NA   NA   NA
#>   Tutume (%)            375   NA  389   NA  403   NA   NA   NA   NA   NA   NA
#>   Kweneng West (%)      152   NA  203   NA  157   NA   NA   NA   NA   NA   NA
#>   Goodhope (%)          196   NA  185   NA  194   NA   NA   NA   NA   NA   NA
#>   Okavango (%)          254   NA  217   NA  327   NA   NA   NA   NA   NA   NA
#>   Hukuntsi (%)          101   NA   88   NA  118   NA   NA   NA   NA   NA   NA
#>   Ngami (%)             418   NA  525   NA  135   NA   NA   NA   NA   NA   NA
#>   Kgatleng (%)          405   NA  450   NA  330   NA   NA   NA   NA   NA   NA
#>   Mabutsane (%)         114   NA  300   NA  300   NA   NA   NA   NA   NA   NA
#>                       year
#> site                   2018 2019 2020 2021 2022 2023 2024 2025
#>   North East (%)         NA   NA   NA   NA   NA   NA   NA   NA
#>   Bobirwa district (%)   NA   NA   NA   NA   NA   NA   NA   NA
#>   Boteti (%)             NA   NA   NA   NA   NA   NA   NA   NA
#>   Gantsi (%)             NA   NA   NA   NA   NA   NA   NA   NA
#>   Chobe (%)              NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgalagadi (%)          NA   NA   NA   NA   NA   NA   NA   NA
#>   Tutume (%)             NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng West (%)       NA   NA   NA   NA   NA   NA   NA   NA
#>   Goodhope (%)           NA   NA   NA   NA   NA   NA   NA   NA
#>   Okavango (%)           NA   NA   NA   NA   NA   NA   NA   NA
#>   Hukuntsi (%)           NA   NA   NA   NA   NA   NA   NA   NA
#>   Ngami (%)              NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgatleng (%)           NA   NA   NA   NA   NA   NA   NA   NA
#>   Mabutsane (%)          NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $Rural$ancrtsite.prev
#>                       year
#> site                   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995
#>   North East (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Bobirwa district (%)   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Boteti (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Gantsi (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Chobe (%)              NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgalagadi (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Tutume (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng West (%)       NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Goodhope (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Okavango (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Hukuntsi (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Ngami (%)              NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgatleng (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Mabutsane (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>                       year
#> site                   1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006
#>   North East (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Bobirwa district (%)   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Boteti (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Gantsi (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Chobe (%)              NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgalagadi (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Tutume (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng West (%)       NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Goodhope (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Okavango (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Hukuntsi (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Ngami (%)              NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgatleng (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Mabutsane (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>                       year
#> site                   2007 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017
#>   North East (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Bobirwa district (%)   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Boteti (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Gantsi (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Chobe (%)              NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgalagadi (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Tutume (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng West (%)       NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Goodhope (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Okavango (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Hukuntsi (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Ngami (%)              NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgatleng (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Mabutsane (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>                       year
#> site                   2018 2019 2020 2021 2022 2023 2024 2025
#>   North East (%)         NA   NA   NA   NA   NA   NA   NA   NA
#>   Bobirwa district (%)   NA   NA   NA   NA   NA   NA   NA   NA
#>   Boteti (%)             NA   NA   NA   NA   NA   NA   NA   NA
#>   Gantsi (%)             NA   NA   NA   NA   NA   NA   NA   NA
#>   Chobe (%)              NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgalagadi (%)          NA   NA   NA   NA   NA   NA   NA   NA
#>   Tutume (%)             NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng West (%)       NA   NA   NA   NA   NA   NA   NA   NA
#>   Goodhope (%)           NA   NA   NA   NA   NA   NA   NA   NA
#>   Okavango (%)           NA   NA   NA   NA   NA   NA   NA   NA
#>   Hukuntsi (%)           NA   NA   NA   NA   NA   NA   NA   NA
#>   Ngami (%)              NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgatleng (%)           NA   NA   NA   NA   NA   NA   NA   NA
#>   Mabutsane (%)          NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $Rural$ancrtsite.n
#>                       year
#> site                   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995
#>   North East (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Bobirwa district (%)   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Boteti (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Gantsi (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Chobe (%)              NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgalagadi (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Tutume (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng West (%)       NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Goodhope (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Okavango (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Hukuntsi (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Ngami (%)              NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgatleng (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Mabutsane (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>                       year
#> site                   1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006
#>   North East (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Bobirwa district (%)   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Boteti (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Gantsi (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Chobe (%)              NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgalagadi (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Tutume (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng West (%)       NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Goodhope (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Okavango (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Hukuntsi (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Ngami (%)              NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgatleng (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Mabutsane (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>                       year
#> site                   2007 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017
#>   North East (%)         NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Bobirwa district (%)   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Boteti (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Gantsi (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Chobe (%)              NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgalagadi (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Tutume (%)             NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng West (%)       NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Goodhope (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Okavango (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Hukuntsi (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Ngami (%)              NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgatleng (%)           NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>   Mabutsane (%)          NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>                       year
#> site                   2018 2019 2020 2021 2022 2023 2024 2025
#>   North East (%)         NA   NA   NA   NA   NA   NA   NA   NA
#>   Bobirwa district (%)   NA   NA   NA   NA   NA   NA   NA   NA
#>   Boteti (%)             NA   NA   NA   NA   NA   NA   NA   NA
#>   Gantsi (%)             NA   NA   NA   NA   NA   NA   NA   NA
#>   Chobe (%)              NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgalagadi (%)          NA   NA   NA   NA   NA   NA   NA   NA
#>   Tutume (%)             NA   NA   NA   NA   NA   NA   NA   NA
#>   Kweneng West (%)       NA   NA   NA   NA   NA   NA   NA   NA
#>   Goodhope (%)           NA   NA   NA   NA   NA   NA   NA   NA
#>   Okavango (%)           NA   NA   NA   NA   NA   NA   NA   NA
#>   Hukuntsi (%)           NA   NA   NA   NA   NA   NA   NA   NA
#>   Ngami (%)              NA   NA   NA   NA   NA   NA   NA   NA
#>   Kgatleng (%)           NA   NA   NA   NA   NA   NA   NA   NA
#>   Mabutsane (%)          NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $Rural$ancrtcens
#>      year   prev     n
#> 2014 2014 0.3060 15765
#> 2015 2015 0.3060 14657
#> 2016 2016 0.2880 13790
#> 2017 2017 0.3005 16430
#> 
#> $Rural$hhs
#>   year  prev    se  n used incid incid_se prev_incid_corr incid_cohort
#> 1 2008 0.281 0.016 NA TRUE    NA       NA              NA           NA
#> 2 2004 0.264 0.014 NA TRUE    NA       NA              NA           NA
#> 3 2012 0.224 0.010 NA TRUE    NA       NA              NA           NA
#> 
#> 
#> attr(,"country")
#> [1] "Botswana"
#> attr(,"country_code")
#> [1] 72
#> attr(,"class")
#> [1] "eppd"