Read best fitting HIV incidence and prevalence data from SPT file.

Locates SPT file within the PJNZ file and reads out the best fitting incidence and prevalence data into memory.

read_spt(pjnz_path)

Arguments

pjnz_path
Path to the PJNZ file.

Value

List of data frames containing the best fit prevalence and incidence data.

Details

Reads best fitting prevalence and incidence and population data for all regions. The outputs one data frame for each region in the SPT file.

Examples

pjnz_path <- system.file("testdata", "Botswana2018.PJNZ", package = "specio")
read_spt(pjnz_path)

#> $National
#>           prev     incid
#> 1970 0.0000000 0.0000000
#> 1971 0.0000000 0.0000000
#> 1972 0.0000000 0.0000000
#> 1973 0.0000000 0.0000000
#> 1974 0.0000000 0.0000000
#> 1975 0.0000014 0.0000014
#> 1976 0.0000033 0.0000025
#> 1977 0.0000088 0.0000065
#> 1978 0.0000226 0.0000162
#> 1979 0.0000554 0.0000383
#> 1980 0.0001295 0.0000875
#> 1981 0.0002952 0.0001916
#> 1982 0.0006428 0.0004002
#> 1983 0.0013318 0.0008034
#> 1984 0.0026699 0.0015414
#> 1985 0.0050887 0.0028098
#> 1986 0.0093220 0.0048889
#> 1987 0.0162066 0.0080584
#> 1988 0.0267099 0.0125413
#> 1989 0.0417915 0.0183906
#> 1990 0.0620990 0.0253377
#> 1991 0.0869335 0.0327951
#> 1992 0.1156197 0.0399857
#> 1993 0.1463996 0.0461151
#> 1994 0.1769462 0.0503236
#> 1995 0.2051639 0.0521529
#> 1996 0.2293791 0.0515567
#> 1997 0.2482859 0.0488538
#> 1998 0.2612451 0.0446445
#> 1999 0.2685677 0.0395861
#> 2000 0.2707136 0.0344049
#> 2001 0.2686649 0.0296824
#> 2002 0.2636959 0.0256882
#> 2003 0.2570185 0.0224640
#> 2004 0.2505687 0.0199410
#> 2005 0.2455405 0.0180344
#> 2006 0.2416744 0.0168564
#> 2007 0.2385946 0.0162537
#> 2008 0.2360219 0.0161560
#> 2009 0.2339781 0.0160109
#> 2010 0.2326550 0.0159864
#> 2011 0.2321532 0.0162477
#> 2012 0.2324049 0.0165680
#> 2013 0.2325216 0.0163815
#> 2014 0.2320371 0.0164349
#> 2015 0.2307951 0.0161874
#> 2016 0.2291630 0.0155479
#> 2017 0.2278076 0.0148740
#> 2018 0.2258526 0.0132630
#> 2019 0.2221546 0.0111814
#> 2020 0.2170800 0.0100824
#> 2021 0.2116689 0.0091834
#> 2022 0.2059472 0.0083845
#> 2023 0.2004120 0.0080525
#> 2024 0.1951668 0.0078788
#> 2025 0.1904657 0.0076460
#> 
#> $Urban
#>           prev     incid    pop
#> 1970 0.0000000 0.0000000  23306
#> 1971 0.0000000 0.0000000  26769
#> 1972 0.0000000 0.0000000  30467
#> 1973 0.0000000 0.0000000  34429
#> 1974 0.0000000 0.0000000  38698
#> 1975 0.0000029 0.0000032  43307
#> 1976 0.0000070 0.0000056  48666
#> 1977 0.0000181 0.0000139  54452
#> 1978 0.0000447 0.0000335  60616
#> 1979 0.0001064 0.0000765  67091
#> 1980 0.0002422 0.0001665  73868
#> 1981 0.0005191 0.0003596  86430
#> 1982 0.0010703 0.0007041  99589
#> 1983 0.0021141 0.0013264 113372
#> 1984 0.0040269 0.0024077 127842
#> 1985 0.0073396 0.0041532 143009
#> 1986 0.0127468 0.0069696 165032
#> 1987 0.0210518 0.0108090 188517
#> 1988 0.0332160 0.0158722 213611
#> 1989 0.0498106 0.0221596 240471
#> 1990 0.0710004 0.0293108 269197
#> 1991 0.0965952 0.0357829 289012
#> 1992 0.1255899 0.0425038 309901
#> 1993 0.1554911 0.0478698 331766
#> 1994 0.1844353 0.0513230 354477
#> 1995 0.2107100 0.0524416 377894
#> 1996 0.2331273 0.0509327 397261
#> 1997 0.2502663 0.0479521 416714
#> 1998 0.2615437 0.0436416 435840
#> 1999 0.2675109 0.0386266 454248
#> 2000 0.2685458 0.0334911 471762
#> 2001 0.2657100 0.0287213 484232
#> 2002 0.2601027 0.0247977 496001
#> 2003 0.2529287 0.0216001 507787
#> 2004 0.2459198 0.0190052 520468
#> 2005 0.2402801 0.0170416 534575
#> 2006 0.2357813 0.0157676 547996
#> 2007 0.2320292 0.0150572 561816
#> 2008 0.2286726 0.0148493 575748
#> 2009 0.2258725 0.0146153 589836
#> 2010 0.2237793 0.0145111 604171
#> 2011 0.2224753 0.0147048 618876
#> 2012 0.2218343 0.0149794 633655
#> 2013 0.2213243 0.0148814 648106
#> 2014 0.2202939 0.0150012 661645
#> 2015 0.2186072 0.0148927 674210
#> 2016 0.2167752 0.0144767 686555
#> 2017 0.2153567 0.0140211 699560
#> 2018 0.2135756 0.0125953 713230
#> 2019 0.2100168 0.0103942 727345
#> 2020 0.2050573 0.0093403 741332
#> 2021 0.1998777 0.0085170 755027
#> 2022 0.1944375 0.0077718 768540
#> 2023 0.1891236 0.0074335 781893
#> 2024 0.1840746 0.0072965 795025
#> 2025 0.1795887 0.0070731 807880
#> 
#> $Rural
#>           prev     incid    pop
#> 1970 0.0000000 0.0000000 275493
#> 1971 0.0000000 0.0000000 283782
#> 1972 0.0000000 0.0000000 292274
#> 1973 0.0000000 0.0000000 301138
#> 1974 0.0000000 0.0000000 310558
#> 1975 0.0000012 0.0000012 320614
#> 1976 0.0000028 0.0000021 330946
#> 1977 0.0000074 0.0000054 341849
#> 1978 0.0000188 0.0000135 352865
#> 1979 0.0000460 0.0000317 363534
#> 1980 0.0001073 0.0000729 373821
#> 1981 0.0002442 0.0001584 379761
#> 1982 0.0005320 0.0003311 384342
#> 1983 0.0011032 0.0006680 387869
#> 1984 0.0022258 0.0012884 390639
#> 1985 0.0042690 0.0023710 392707
#> 1986 0.0078730 0.0041336 390044
#> 1987 0.0138453 0.0069003 386821
#> 1988 0.0230820 0.0109299 383082
#> 1989 0.0367012 0.0163108 378832
#> 1990 0.0556909 0.0228500 373941
#> 1991 0.0795668 0.0306791 379047
#> 1992 0.1075725 0.0381011 383953
#> 1993 0.1386350 0.0447275 388460
#> 1994 0.1701802 0.0494868 392360
#> 1995 0.1998643 0.0518965 395473
#> 1996 0.2256782 0.0521448 402325
#> 1997 0.2462651 0.0497356 408381
#> 1998 0.2609302 0.0456626 413262
#> 1999 0.2697197 0.0405972 416668
#> 2000 0.2731571 0.0354041 418531
#> 2001 0.2720461 0.0307726 423181
#> 2002 0.2678705 0.0267161 426921
#> 2003 0.2618437 0.0234783 430399
#> 2004 0.2561389 0.0210584 434389
#> 2005 0.2519413 0.0192404 439334
#> 2006 0.2489094 0.0182018 446359
#> 2007 0.2467275 0.0177483 453534
#> 2008 0.2452077 0.0178063 460643
#> 2009 0.2442012 0.0177935 467662
#> 2010 0.2439514 0.0178923 474703
#> 2011 0.2445930 0.0182638 481477
#> 2012 0.2461272 0.0186697 488118
#> 2013 0.2472038 0.0183915 494278
#> 2014 0.2475896 0.0183794 499589
#> 2015 0.2470995 0.0179644 503985
#> 2016 0.2459436 0.0170351 506831
#> 2017 0.2448840 0.0160739 510067
#> 2018 0.2429063 0.0142145 513454
#> 2019 0.2392339 0.0123171 516904
#> 2020 0.2342201 0.0111667 519999
#> 2021 0.2287146 0.0101697 522286
#> 2022 0.2228197 0.0093035 524268
#> 2023 0.2171932 0.0089929 525969
#> 2024 0.2118890 0.0087754 527356
#> 2025 0.2070956 0.0085402 528401
#>